% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExQA.R
\docType{data}
\name{ExQA}
\alias{ExQA}
\title{Landsat 8 QA_PIXEL data from San Clemente Island, CA 2015-11-27}
\format{
A `raster` class object.
}
\usage{
data(ExQA)
}
\description{
A raster representing Lansat 8 (OLI/TIRS) Collection 2 Tier 1 QA_PIXEL data associated
with imagery taken on November 27, 2015.
The data has been cropped to the extent 352244, 3648387:356562, 3653420 (EPSG:32611).
}
\keyword{datasets}
