% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExRed.R
\docType{data}
\name{ExRed}
\alias{ExRed}
\title{Landsat 8 Band 4 Imagery from San Clemente Island, CA 2015-11-27}
\format{
A `raster` class object.
}
\usage{
data(ExRed)
}
\description{
A raster representing Lansat 8 (OLI/TIRS) Collection 2 Tier 1 Band 4 data taken on November 27, 2015.
The data has been cropped to the extent 352244, 3648387:356562, 3653420 (EPSG:32611).
}
\keyword{datasets}
