% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetAccess.R
\name{GetAccess}
\alias{GetAccess}
\title{Read a MS Access table as a data table}
\usage{
GetAccess(SheetName, Path)
}
\arguments{
\item{SheetName}{A character string matching the name of the MS Access table to be read.}

\item{Path}{A character string containing the path to the MS Access database within which the table
of interest resides.}
}
\value{
This function returns the table of interest (specified using the \code{SheetName} argument) as a
data table.
}
\description{
This function reads a specified MS Access table as a data table. In a Windows OS, it requires MS Access drivers which can
currently be found \href{https://www.microsoft.com/en-us/download/details.aspx?id=54920}{here}. If using
this function causes R to crash while copying and pasting or switching tabs, consider installing Access
Runtime which can be found \href{https://www.microsoft.com/en-us/download/details.aspx?id=50040}{here}.
In a Linux OS, it requires 'mdbtools' which can be installed in the terminal with "sudo apt install mdbtools".
}
\examples{
\dontrun{
DbPath <- "C:/Users/Hunter/Documents/ExampleDb.accdb"

GetAccess("Example Access Table Name", DbPath)
}
}
