% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LandsatBinaryCodes.R
\docType{data}
\name{LandsatBinaryCodes}
\alias{LandsatBinaryCodes}
\title{All Landsat QA_PIXEL values and associated parsed binary values}
\format{
A `data.table` class object.
}
\usage{
data(LandsatBinaryCodes)
}
\description{
A data table showing all possible 16 bit unsigned QA_PIXEL (Landsat imagery quality assurance) raster values
and parsed underlying binary code that provides quality information about Landsat imagery pixels. Additional
information pertaining to QA_PIXEL codes and their meaning can be found \href{https://d9-wret.s3.us-west-2.amazonaws.com/assets/palladium/production/s3fs-public/atoms/files/LSDS-1328_Landsat8-9-OLI-TIRS-C2-L2-DFCB-v6.pdf}{here}.
}
\keyword{datasets}
