% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LCP_Calc.R
\name{LCP_Calc}
\alias{LCP_Calc}
\title{Calculate a least-cost path and write to file}
\usage{
LCP_Calc(
  StartE,
  StartN,
  EndE,
  EndN,
  Lcoord,
  Tcoord,
  Rcoord,
  Bcoord,
  ResampleFactor = NA,
  DtmPath,
  SlopePath,
  SaveDir,
  FileName,
  proj4str,
  SteepAngle = 99,
  CostFun = "ree"
)
}
\arguments{
\item{StartE, StartN}{Numeric values indicating the starting longitude and latitude provided in the same
coordinate system as the DEM and slope layers.}

\item{EndE, EndN}{Numeric values indicating the ending longitude and latitude provided in the same
coordinate system as the DEM and slope layers.}

\item{Lcoord, Tcoord, Rcoord, Bcoord}{Numeric values indicating maximum extent longitudes and
latitudes to consider for the LCP. The \code{Lcoord}, \code{Tcoord}, \code{Rcoord} and \code{Bcoord}
arguments are associated with the west-most longitude, north-most latitude, east-most longitude and
south-most latitude to consider, respectively.}

\item{ResampleFactor}{A numeric value or \code{NA} used to specify the factor by which the DEM layer's
cells should be increased by for resampling. This is meant to reduce computational
load. If no resampling is required, use \code{NA}. For most accurate results, use the lowest resampling
factor possible, or no resampling factor \code{NA}.}

\item{DtmPath}{A character string supplying the path to the DEM layer.}

\item{SlopePath}{A character string supplying the path to the slope layer using degrees for slope units,
and using the same CRS as the DEM layer.}

\item{SaveDir}{A character string supplying the path to the directory the LCP should be saved to.}

\item{FileName}{A character string supplying the base file name for LCP shapefile components.}

\item{proj4str}{A character string supplying the proj4 string for the slope and DEM layers' CRS.}

\item{SteepAngle}{A numeric value indicating the minimum angle (in degrees) that should be considered
prohibitively steep. If no terrain is to be considered prohibitively steep, use a value >90.}

\item{CostFun}{A character string supplying the cost function to be used. For details about cost function
options, see \code{\link[movecost]{movecorr}}.}
}
\value{
This function returns a shapefile representing the calculated LCP that can be read into R or
a GIS program (e.g., QGIS, ArcMap).
}
\description{
This function uses the \code{\link[movecorr]{movecost}} function to calculate least-cost paths (LCPs)
using user supplied DEM and slope layers. Users can also specify what they would like to consider a
prohibitively steep slope to try and avoid steep terrain. This function is not perfect, should be used
only as a tool for considering potential movement paths, and should always be examined by someone with
terrain navigation experience if used for route planning.
}
\examples{
\dontrun{
DemPath <- "C:/Users/Hunter/SciDem.tif"
Slope <- "C:/Users/Hunter/SciSlope.tif"

SaveFolder <- "C:/Users/Hunter/Lcp"

LCP_Calc(365556, 3639062,
         365644, 3638898,
         364974, 3639623,
         365868, 3638111,
         ResampleFactor = 1.5,
         DtmPath = DemPath,
         SlopePath = Slope,
         SaveDir = SaveFolder,
         FileName = "TestLCP",
         SteepAngle = 60)
}

}
