% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ListTables.R
\name{ListTables}
\alias{ListTables}
\title{List tables and queries in an Access database}
\usage{
ListTables(Path, Pattern = NULL)
}
\arguments{
\item{Path}{A character string containing the path to an MS Access database.}

\item{Pattern}{An optional regular expression used to search for table/query names.}
}
\value{
This function returns the names of all tables and queries if no pattern is
specified, or the names of tables and queries that match a specified pattern.
}
\description{
This function returns all tables and queries in a specified MS Access database.
}
\examples{
\dontrun{
DbPath <- "C:/Users/Hunter/Documents/ExampleDb.accdb"

# Return all table and query names
ListTables(DbPath)

# Return only tbale and query names that contain the word "Rat"
ListTables(DbPath, "Rat")
}
}
