% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotDate.R
\name{PlotDate}
\alias{PlotDate}
\title{A 'Round Down' Function for Date Class Values}
\usage{
PlotDate(x, Aggregate_Unit = "months")
}
\arguments{
\item{x}{A Date class value or vector.}

\item{Aggregate_Unit}{The temporal unit that should be used for rounding date values down.
Currently the only acceptable values for this argument are "months" and "years".}
}
\description{
Use this function to round Date class values down to the first date of the month or year.
}
\examples{
library(lubridate)

TestVector <- ymd(c("2022-01-03", "2023-12-24"))

PlotDate(TestVector)

PlotDate(TestVector, Aggregate_Unit = "years")
}
