% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_PREM.R
\name{theme_PREM}
\alias{theme_PREM}
\title{A simple and elegant ggplot2 theme}
\usage{
theme_PREM()
}
\description{
A pre-packaged theme. To make inline modifications to this theme, use \code{\link[ggplot2]{theme}}.
}
\examples{
#Creating a plot with theme_nocturnal
library(ggplot2)

Plot <- ggplot(cars, aes(speed, dist)) +
  geom_point() +
  theme_PREM()

#Changing the axis text size of a theme_PREM plot

Plot2 <- Plot +
  theme(axis.text = element_text(size = 4))
}
