% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TimeChr.R
\name{TimeChr}
\alias{TimeChr}
\title{Extract a hh:mm:ss time character string with leading zeros from a POSIX class object}
\usage{
TimeChr(DateTime)
}
\arguments{
\item{DateTime}{A POSIX class value or vector}
}
\value{
This function returns a hh:mm:ss time character vector of the same length as `DateTime`.
}
\description{
This function extracts the hour, minute and second values from a PSOIX class object, separates these
values with colons, and adds leading zeros as necessary.
}
\examples{
TimeChr(Sys.time())

}
