% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WKT_Trunc.R
\name{WKT_Trunc}
\alias{WKT_Trunc}
\title{Truncate WKT-formatted tabular data decimal points}
\usage{
WKT_Trunc(File, OutputFile, Digits = 1, Separator = ";")
}
\arguments{
\item{File}{A character string supplying the path to a tabular (e.g., CSV, TXT) WKT-formatted
file providing vertices for a point, line or polygon layer.}

\item{OutputFile}{A character string supplying the full path to save the output file to.}

\item{Digits}{An integer indicating the number of decimal places to maintain.}

\item{Separator}{A character string indicating the column separator character used in the input file.}
}
\value{
A tabular file with WKT-formatted coordinates containing the specified number of decimal places.
}
\description{
This function decreases the number of decimal points associated with WKT coordinate strings.
}
\examples{
\dontrun{
TestFile <- "C:/Users/Hunter/Documents/TestWKT.csv"
TestOutput <- "C:/Users/Hunter/Documents/OutputWKT.csv"

WKT_Trunc(TestFile, TestOutput, 2, Separator = ",")
}
}
