% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_PREM.R
\name{theme_PREM}
\alias{theme_PREM}
\title{A simple and elegant ggplot2 theme}
\usage{
theme_PREM(MajorGrid = "white", MinorGrid = "white")
}
\arguments{
\item{MajorGrid}{A character string specifying the color of major grid lines.}

\item{MinorGrid}{A character string specifying the color of minor grid lines.}
}
\description{
A pre-packaged theme. To make inline modifications to this theme, use \code{\link[ggplot2]{theme}}.
}
\examples{
#Creating a plot with theme_PREM

library(ggplot2)

Plot <- ggplot(cars, aes(speed, dist)) +
  geom_point() +
  theme_PREM()

#Creating a plot with visible grid lines

PlotLines <- ggplot(cars, aes(speed, dist)) +
  geom_point() +
  theme_PREM(MajorGrid = "grey90", MinorGrid = "grey95")

#Changing the axis text size of a theme_PREM plot

Plot2 <- Plot +
  theme(axis.text = element_text(size = 4))
}
